ocr-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
Pillow==10.1.0
rapidocr-onnxruntime
pdf2image==1.16.3
pypdfium2==4.26.0
